Move BlockId to common
