+++ common/src/main/java/tachyon/master/block/BlockId.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.master.next.block.meta;
+package tachyon.master.block;
 
 // TODO: this should probably be moved to common/.
 public class BlockId {
