+++ servers/src/main/java/tachyon/master/next/block/BlockMaster.java
@@ -32,9 +32,9 @@ import org.slf4j.LoggerFactory;
 import tachyon.Constants;
 import tachyon.StorageDirId;
 import tachyon.master.next.Master;
-import tachyon.master.next.block.meta.MasterWorkerInfo;
 import tachyon.master.next.block.meta.MasterBlockInfo;
 import tachyon.master.next.block.meta.MasterBlockLocation;
+import tachyon.master.next.block.meta.MasterWorkerInfo;
 import tachyon.thrift.BlockInfo;
 import tachyon.thrift.BlockLocation;
 import tachyon.thrift.Command;
