+++ servers/src/main/java/tachyon/master/next/filesystem/FileSystemMaster.java
@@ -31,9 +31,9 @@ import tachyon.PrefixList;
 import tachyon.TachyonURI;
 import tachyon.conf.TachyonConf;
 import tachyon.master.Dependency;
+import tachyon.master.block.BlockId;
 import tachyon.master.next.Master;
 import tachyon.master.next.block.BlockMaster;
-import tachyon.master.next.block.meta.BlockId;
 import tachyon.master.next.block.meta.MasterWorkerInfo;
 import tachyon.master.next.filesystem.meta.DependencyMap;
 import tachyon.master.next.filesystem.meta.Inode;
