+++ servers/src/main/java/tachyon/master/next/filesystem/meta/InodeTree.java
@@ -30,7 +30,7 @@ import com.google.common.collect.Sets;
 import tachyon.Constants;
 import tachyon.Pair;
 import tachyon.TachyonURI;
-import tachyon.master.next.block.meta.BlockId;
+import tachyon.master.block.BlockId;
 import tachyon.master.next.block.ContainerIdGenerator;
 import tachyon.thrift.BlockInfoException;
 import tachyon.thrift.FileAlreadyExistException;
